wifi: Fix MQTT never connecting when using static IPS.
The MQTT module never connected to a broker if static IPs where used,
because the WIFI module never sets the state to ::CONNECTED.

For some reason SYSTEM_EVENT_STA_CONNECTED did not fire in this case,
but instead SYSTEM_EVENT_STA_GOT_IP fired twice. This seems to be
a known problem:
https://github.com/espressif/arduino-esp32/issues/2856

We now set the ::CONNECTED state also in the GOT_IP event to
circumvent this.